Demo/Monitoring/requirements.txt
# Core ML & data
scikit-learn==1.3.2
pandas==2.1.1
joblib==1.3.2
numpy==1.27.5

# HTTP / requests
requests==2.32.1

# YARA
yara-python==4.3.0

# Pyparsing (needed for YAML & sigma YAML parsing)
pyparsing==3.1.1

# YAML parsing (for Sigma rules)
PyYAML==6.0

# Optional: color printing
colorama==0.4.6

# SQLite comes with Python, no need to include
